issue #11620: pymorphy2_lookup, pymorphy2 lemmatizer refactoring
